+++ tests/conftest.py
@@ -1,10 +1,34 @@
 # -*- coding: utf-8 -*-
-"""
-    Dummy conftest.py for mesi_search.
+import pytest
 
-    If you don't know what this is for, just leave it empty.
-    Read more about conftest.py under:
-    https://pytest.org/latest/plugins.html
-"""
 
-# import pytest
+@pytest.fixture(scope="module")
+def candig_raw_results():
+    test_data = {
+        "dataset-1": {
+            "results": {
+                "patients": [
+                    {
+                        "causeOfDeath": {
+                            "Cancer": 32,
+                            "Heart": 23,
+                        }
+                    }
+                ]
+            }
+        },
+
+        "dataset-2": {
+            "results": {
+                "patients": [
+                    {
+                        "causeOfDeath": {
+                            "Cancer": 11,
+                            "Heart": 33,
+                        }
+                    }
+                ]
+            }
+        }
+    }
+    return test_data