+++ tests/test_test.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-
-import pytest
-
-
-def test_test():
-    assert 1 == 1