+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+### Added
+- Tests (2020-08-13)
+- Counts based approach instead of sum/percentage (2020-08-24)
+- Session based tracking of requests per user (2020-08-25)
+
 ## 0.1.0 - 2020-08-11
 ### Fixed
 - `Sphinx` docs pointing to incorrect `rst` files